Refactoringr 1
